addressed the  ruff type errors.
I change certain types to make the expectation compatible with the returned value.

+++ tests/test_trace.py
@@ -51,7 +51,7 @@ def test_trace_pythia() -> None:
     p = np.genfromtxt(
         "test_data/trace_csvs/selection0.csv",
         delimiter=",",
-        dtype=np.int_,
+        dtype=np.double,
     )
     p = p - 1  # Adjusting indices to be zero-based
 
@@ -67,7 +67,7 @@ def test_trace_pythia() -> None:
 
     # Initialising and running the TRACE analysis
     trace = Trace()
-    trace.run(z, y_bin, p, beta, algo_labels, trace_options)
+    trace.run(z, y_bin, p.astype(np.double), beta, algo_labels, trace_options)
 
 
 def test_trace_simulation() -> None:
@@ -106,7 +106,7 @@ def test_trace_simulation() -> None:
     p2 = np.genfromtxt(
         "test_data/trace_csvs/dataP.csv",
         delimiter=",",
-        dtype=np.int_,
+        dtype=np.double,
     )
     p2 = p2 - 1  # Adjusting indices to be zero-based
 
@@ -122,4 +122,4 @@ def test_trace_simulation() -> None:
 
     # Initialising and running the TRACE analysis
     trace = Trace()
-    trace.run(z, y_bin2, p2, beta, algo_labels, trace_options)
+    trace.run(z, y_bin2, p2.astype(np.double), beta, algo_labels, trace_options)
